shims: move around the memory shim / add file functions
