+++ shims/src/memory/mod.rs
@@ -1,0 +1,11 @@
+//! Memory shims.
+//! (TODO)
+
+pub mod error;
+
+mod simple;
+mod file_backed;
+
+pub use simple::MemoryShim;
+pub use file_backed::FileBackedMemoryShim;
+
